Create Grouping Data
Task:
1. Let's apply grouping to our DataFrame and see how it works with the count() and sum() methods. Grouping our DataFrame by the user_id column, we will find out two things about our users: 1) how much time each of them spent listening to music, and 2) how many songs each of them listened to. We will print two objects: a Series showing the sums of total_play_seconds values for each user, and a new DataFrameGroupBy showing the number of entries across all columns for each user.
To do so, first group the DataFrame by the user_id column and apply the sum() method to one column: total_play_seconds . Store this calculation in the variable user_playtime . Next, group the DataFrame again by user_id and apply the count() method without specifying a column. Store it in the variable user_plays. We don't have missing values, so the values across all columns for each group should be the same, and count() will give us the number of rows corresponding to each of the users. In this case, the number of rows will equal the number of listens. Print the first five rows of each object.

2. Let's see if we have users with more than 50 song listens a day. To find out, first we need to devise a general tool that we can apply to our situation.
Write a function, get_heavy_listener, which accepts a certain grouping as its groups argument. The function should sift through the entire grouping.
There are two elements in each group: a group name and a a selected part of the original DataFrame. You'll see why we need groups in the next task
Here, we want our function to loop through a group and do something with two elements each time. To do so, we'll write for group_name, group_data in groups —that means that we expect the passed value (group) to have two values inside. We use the arbitrary variable names group_name (in the next task, this variable will represent users) and group_data to address them.
When the function finds a group in which the selected part of the original DataFrame contains more than 50 rows (i.e. the length of this part of the data is greater than 50; in our case, that would mean that a certain user had more than 50 listens), it returns the group name. After it finds one such group, it terminates without looking any further.
In this task, we will have no output, but the checker will verify if your code is correct.

3. Now you'll see why we needed groups: to look for specific values in a particular part of the DataFrame. We'll use our new function, but first we need to think of an argument. Let's group the DataFrame by the user_id column, so group_name will represent the user and the corresponding piece of data will be handled by the group_data variable. Now that we've decided what data to use as the argument, we can save it to user_grouping .
Call the get_heavy_listener function and pass user_grouping as its argument. Save the result in the user_id variable and print it. Note that the algorithm is not perfect: as soon as it finds a user with 50+ listens, it stops searching. But here we have at least one of our unknown music lovers!