Create Sorting Data
Task:

1. You have discovered a music lover in the Yandex.Music table with unique data. This user has listened to more than 50 songs in a single day.
Acquire a table with the songs they listened to. In order to do this, request rows from the df data structure that meet two conditions simultaneously:
The value in the ‘user_id’ column must be equal to the value of the user_id variable.
The listening time (the value in the ‘total_play_seconds’ column) cannot be less or equal to 0. Save the results to the variable music_user.

2. Now we’ll find out how much time this user has spent listening to music of each genre.
Group the data of the music_user table by the ‘genre_name’ column, and sum the values in the column ‘total_play_seconds’. Save the results to the variable sum_music_user and print its value.

3. It looks like our music lover’s preferences are beginning to emerge; however, song lengths can vary greatly from genre to genre. It’s important for us to figure out how many songs from each genre the user has played. Group the data of the music_user table by the genre_name column, then calculate how many values are in the genre_name column. Save the results to the count_music_user variable and print its value.
It's important to specify the column as genre_name in the square brackets before .count(). If you don't specify the column, the resulting table will have several columns and the platform will not accept your solution. Although we know there are no missing values in the table, and the output will be the same for any column, that's not always going to be the case. So let's practice specifying one column for count(): genre_name.
Comment out the “print sum” command from the previous task so that it doesn’t get in the way of your new results.

4. We need to position the largest values at the top so the user’s preferences are readily apparent. Sort the data in the sum_music_user grouping in descending order. Caution: when applying the sort_values() method to a Series with a single column, you don’t need the by argument, only the sorting order.
Save the result to the variable final_sum and print its value.
Comment out the print() commands from the previous tasks.

5. Now we will have to do the same with the number of songs the music lover has listened to. Sort the data from the count_music_user grouping in descending order. Save the results to the variable final_count and print its value.
Comment out the “print” command from the previous task.