Create Solving Problems and Presenting Results
Task:

1. It's a big moment! We're about to compile a DataFrame displaying how metrics change over time for a given genre. We have two datasets: an old one and a new one. The paths to these datasets are already in the precode. We will find the total number of plays and median duration for pop songs in both datasets.
Following the instructions in the precode, set the genre variable to 'Pop'. There is a loop in the precode that will process both the old and updated DataFrames.
Create a temporary storage list temp in the first line of the loop. The data inside it will comprise a future row of the final DataFrame.
In the second line, create a selection based on two conditions (with &). genre_name should be equal to the genre variable, and total_play_seconds should be above zero.
In the third line, count the user_id column of the selection and add it to temp using append.
Then calculate and add the median for total_play_seconds, and again add it to temp using append.
Finally, add the contents of temp to the data list.
Now let's create the final DataFrame. To do so, use the pd.DataFrame method with three parameters: data, index, and columns. (See the precode.)
Notice how we use the variable df to represent a chosen DataFrame inside the cycle and genre to represent the genre. The value for df is set by the loop and genre is set manually. If you want to play around and pick data for another genre, you don't need to find all the lines with 'Pop' throughout the code; just change the variable at the beginning.

2. Finally, calculate the post-experiment happiness metric using the instructions from the lesson. Use two datasets. Make a DataFrame and print it. Make sure to follow the instructions and use the variable names given in the precode.